requirements.txt
pyttsx3>=2.90
speechrecognition>=3.8.1
pyautogui>=0.9.54
pyaudio>=0.2.11
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pywhatkit>=5.4
wikipedia>=1.4.0
wolframalpha>=5.0.0
pyjokes>=0.6.0
psutil>=5.9.0
pillow>=10.0.0
pywin32>=305; sys_platform == "win32"
